+++ .gitignore
@@ -1,5 +1,5 @@
 node_modules
-task-1/txt
+core/txt
 .vscode
 .idea
 .env
