+++ core/task-1.2-read-entire-file-babel.js
