+++ core/task-1.2-use-streams-babel.js
