+++ core/utils/getLowerCaseTransformStream.js
