+++ core/utils/logMemoryUsage-babel.js
