+++ core/utils/logMemoryUsage.js
