+++ package.json
@@ -4,12 +4,12 @@
   "description": "Node.js training course",
   "main": "index.js",
   "scripts": {
-    "task-1.1": "nodemon task-1/task-1.1.js",
-    "task-1.1:babel": "nodemon --exec babel-node task-1/task-1.1-babel.js",
-    "task-1.2": "nodemon task-1/task-1.2-read-entire-file.js",
-    "task-1.2:babel": "nodemon --exec babel-node task-1/task-1.2-read-entire-file-babel.js",
-    "task-1.2-streams": "nodemon task-1/task-1.2-use-streams.js",
-    "task-1.2-streams:babel": "nodemon --exec babel-node task-1/task-1.2-use-streams-babel.js",
+    "task-1.1": "nodemon core/task-1.1.js",
+    "task-1.1:babel": "nodemon --exec babel-node core/task-1.1-babel.js",
+    "task-1.2": "nodemon core/task-1.2-read-entire-file.js",
+    "task-1.2:babel": "nodemon --exec babel-node core/task-1.2-read-entire-file-babel.js",
+    "task-1.2-streams": "nodemon core/task-1.2-use-streams.js",
+    "task-1.2-streams:babel": "nodemon --exec babel-node core/task-1.2-use-streams-babel.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon --exec ts-node src/index.ts",
     "compile": "rm -rf dist && mkdir dist && tsc"
