+++ src/index.ts
@@ -2,14 +2,14 @@ import express from 'express';
 import dotenv from 'dotenv';
 
 import { expressErrorHandler } from './errors/error.handler';
-import { userController } from './user/user.controller';
+import {userController} from "./entities/user/user.route";
 
 dotenv.config();
 
 const app = express();
 const appRouter = express.Router();
 
-app.use(express.json())
+app.use(express.json());
 appRouter.use('/api', [userController]);
 app.use(appRouter);
